updates foe the construction of the app
added click functions on the task, and in the float button. the FB show a create window, and when clicked directly on the task in the list, shows a update task window.

next steps will be when created just add in the tesks and when uptaded the text will be uptaded in the same task id